Move create_bump_LUT to top level korlib
