+++ korlib/bumpmap.h
@@ -1,0 +1,29 @@
+/* This file is part of Korman.
+ *
+ * Korman is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Korman is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Korman.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef _KORLIB_BUMPMAP_H
+#define _KORLIB_BUMPMAP_H
+
+#include "korlib.h"
+
+extern "C" {
+
+PyObject* create_bump_LUT(PyObject*, PyObject* args);
+
+};
+
+#endif // _KORLIB_BUMPMAP_H
+
