+++ korlib/module.cpp
@@ -15,12 +15,14 @@
  */
 
 #include "buffer.h"
+#include "bumpmap.h"
 #include "sound.h"
 #include "texture.h"
 
 extern "C" {
 
 static PyMethodDef korlib_Methods[] = {
+    { _pycs("create_bump_LUT"), (PyCFunction)create_bump_LUT, METH_VARARGS, NULL },
     { _pycs("inspect_vorbisfile"), (PyCFunction)inspect_vorbisfile, METH_VARARGS, NULL },
 
     { NULL, NULL, 0, NULL },
